Move Exception classes, remove CmpException, fix CS issues
