+++ Plugin/Framework/Search/Request/AbstractHandler.php
@@ -38,7 +38,6 @@ namespace Nosto\Cmp\Plugin\Framework\Search\Request;
 
 use Exception;
 use Magento\Store\Model\StoreManagerInterface;
-use Nosto\Cmp\Exception\CmpException;
 use Nosto\Cmp\Helper\Data as CmpHelperData;
 use Nosto\Cmp\Helper\SearchEngine;
 use Nosto\Cmp\Model\Facet\FacetInterface;
@@ -148,7 +147,7 @@ abstract class AbstractHandler
                 $this->parsePageNumber($requestData),
                 $this->parseLimit($requestData)
             );
-        } catch (CmpException $e) {
+        } catch (Exception $e) {
             $this->logger->exception($e);
             return;
         }
@@ -254,14 +253,14 @@ abstract class AbstractHandler
     /**
      * @param array $requestData
      * @return int
-     * @throws CmpException
+     * @throws Exception
      */
     abstract public function parsePageNumber(array $requestData);
 
     /**
      * @param array $requestData
      * @return int
-     * @throws CmpException
+     * @throws Exception
      */
     abstract public function parseLimit(array $requestData);
